+++ src/cmd/gz_TEST.cc
@@ -61,7 +61,7 @@ std::string get_config_path(const std::string filename)
 /////////////////////////////////////////////////
 TEST(CmdLine, Ls)
 {
-  std::string cmd = "ign launch " + get_config_path("ls.ign");
+  std::string cmd = "gz launch " + get_config_path("ls.ign");
   std::string output = customExecStr(cmd);
   EXPECT_TRUE(output.find("CMakeFiles") != std::string::npos) << output;
   EXPECT_TRUE(output.find("Makefile") != std::string::npos) << output;
@@ -71,7 +71,7 @@ TEST(CmdLine, Ls)
 TEST(CmdLine, EchoSelf)
 {
   std::string filePath = get_config_path("echo.ign");
-  std::string cmd = "ign launch " + filePath;
+  std::string cmd = "gz launch " + filePath;
   std::string output = customExecStr(cmd);
   EXPECT_EQ(filePath, output) << output;
 }
@@ -79,7 +79,7 @@ TEST(CmdLine, EchoSelf)
 /////////////////////////////////////////////////
 TEST(CmdLine, HelpSelf)
 {
-  std::string output = customExecStr("ign launch --help");
+  std::string output = customExecStr("gz launch --help");
   EXPECT_NE(std::string::npos,
     output.find("Introspect Ignition launch")) << output;
 }
@@ -88,7 +88,7 @@ TEST(CmdLine, HelpSelf)
 TEST(CmdLine, EchoErb)
 {
   std::string filePath = get_config_path("erb.ign");
-  std::string cmd = "ign launch " + filePath + " testVar:=erb1234";
+  std::string cmd = "gz launch " + filePath + " testVar:=erb1234";
   std::string output = customExecStr(cmd);
   EXPECT_EQ("erb1234", output) << output;
 }
@@ -97,7 +97,7 @@ TEST(CmdLine, EchoErb)
 TEST(CmdLine, EchoBadErb)
 {
   std::string filePath = get_config_path("erb.ign");
-  std::string cmd = " ign launch " + filePath + " badargument";
+  std::string cmd = " gz launch " + filePath + " badargument";
   std::string output = customExecStr(cmd);
   EXPECT_NE(std::string::npos, output.find("is wrong for erb")) << output;
 }