+++ Changelog.md
@@ -179,6 +179,15 @@
 
 ### Gazebo Launch 1.X.X (20XX-XX-XX)
 
+### Gazebo Launch 1.10.1 (2020-12-23)
+
+1. Fix codecheck
+   * [Pull Request 67](https://github.com/gazebosim/gz-launch/pull/67)
+   * [Pull Request 70](https://github.com/gazebosim/gz-launch/pull/70)
+
+1. Fix race condition in websocket server
+   * [Pull Request 68](https://github.com/gazebosim/gz-launch/pull/68)
+
 ### Gazebo Launch 1.10.0 (2020-09-25)
 
 * Modernize Github Actions CI.